python/requirements.txt
fastapi
uvicorn
watchfiles
psycopg
psycopg[binary] 
yoyo-migrations
sqlalchemy 
